Android app. Mostly just activities.
Some of this is boiler-plate code generated by Android Studio. A good
bit of this has been modified for our purposes. There is an initial
login screen. There is a functioning Navigation Drawer activity, with
Fragments to show each type of Listing. When clicking on a listing,
there is a detailed Activity for each Listing type.